--- java/src/processing/mode/java/pdex/util/runtime/strategy/ModeSketchRuntimePathFactory.java
@@ -1,50 +1,0 @@
-/* -*- mode: java; c-basic-offset: 2; indent-tabs-mode: nil -*- */
-
-/*
-Part of the Processing project - http://processing.org
-Copyright (c) 2012-19 The Processing Foundation
-
-This program is free software; you can redistribute it and/or modify
-it under the terms of the GNU General Public License version 2
-as published by the Free Software Foundation.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program; if not, write to the Free Software Foundation, Inc.
-51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
-*/
-
-package processing.mode.java.pdex.util.runtime.strategy;
-
-import processing.app.Library;
-import processing.app.Sketch;
-import processing.mode.java.JavaMode;
-import processing.mode.java.pdex.ImportStatement;
-import processing.mode.java.pdex.util.runtime.RuntimePathUtil;
-
-import java.util.ArrayList;
-import java.util.List;
-
-
-/**
- * Runtime path factory for classpath entries required by the processing mode like {JavaMode}.
- */
-public class ModeSketchRuntimePathFactory implements RuntimePathFactoryStrategy {
-
-  @Override
-  public List<String> buildClasspath(JavaMode mode, List<ImportStatement> imports, Sketch sketch) {
-    Library coreLibrary = mode.getCoreLibrary();
-    String coreClassPath = coreLibrary != null ?
-        coreLibrary.getClassPath() : mode.getSearchPath();
-    if (coreClassPath != null) {
-      return RuntimePathUtil.sanitizeClassPath(coreClassPath);
-    } else {
-      return new ArrayList<>();
-    }
-  }
-
-}
