+++ java/test/processing/mode/java/pdex/util/LibrarySketchRuntimePathFactoryTest.java
@@ -18,7 +18,7 @@ along with this program; if not, write to the Free Software Foundation, Inc.
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 */
 
-package processing.mode.java.pdex.util.runtime.strategy;
+package processing.mode.java.pdex.util;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -30,34 +30,35 @@ import java.util.List;
 
 import static org.junit.Assert.*;
 
-public class ModeSketchRuntimePathFactoryTest {
 
-  private ModeSketchRuntimePathFactory factory;
+public class LibrarySketchRuntimePathFactoryTest {
+
+  private RuntimePathBuilder.RuntimePathFactoryStrategy factory;
   private JavaMode testMode;
   private List<ImportStatement> testImports;
   private Sketch testSketch;
-  private List<String> classpath;
+  private List<String> classpathEntries;
 
   @Before
   public void setUp() throws Exception {
-    factory = new ModeSketchRuntimePathFactory();
+    RuntimePathBuilder builder = new RuntimePathBuilder();
+    factory = builder::buildLibrarySketchPath;
     testMode = RuntimePathFactoryTestUtil.createTestJavaMode();
     testImports = RuntimePathFactoryTestUtil.createTestImports();
     testSketch = RuntimePathFactoryTestUtil.createTestSketch();
 
-    classpath = factory.buildClasspath(testMode, testImports, testSketch);
+    classpathEntries = factory.buildClasspath(testMode, testImports, testSketch);
   }
 
   @Test
-  public void buildClasspathLength() {
-    assertEquals(3, classpath.size());
+  public void testBuildClasspathSize() {
+    assertEquals(2, classpathEntries.size());
   }
 
   @Test
-  public void buildClasspathValues() {
-    assertEquals("library6", classpath.get(0));
-    assertEquals("javax.library7", classpath.get(1));
-    assertEquals("library8", classpath.get(2));
+  public void testBuildClasspathValues() {
+    assertTrue(classpathEntries.get(0).contains("library3"));
+    assertTrue(classpathEntries.get(1).contains("library5"));
   }
 
 }
