+++ java/test/processing/mode/java/pdex/util/ModeSketchRuntimePathFactoryTest.java
@@ -18,7 +18,7 @@ along with this program; if not, write to the Free Software Foundation, Inc.
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 */
 
-package processing.mode.java.pdex.util.runtime.strategy;
+package processing.mode.java.pdex.util;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -26,39 +26,40 @@ import processing.app.Sketch;
 import processing.mode.java.JavaMode;
 import processing.mode.java.pdex.ImportStatement;
 
-import java.io.IOException;
 import java.util.List;
 
 import static org.junit.Assert.*;
 
-public class CoreLibraryRuntimePathFactoryTest {
 
-  private CoreLibraryRuntimePathFactory factory;
+public class ModeSketchRuntimePathFactoryTest {
+
+  private RuntimePathBuilder.RuntimePathFactoryStrategy factory;
   private JavaMode testMode;
   private List<ImportStatement> testImports;
   private Sketch testSketch;
-  private List<String> classPath;
+  private List<String> classpath;
 
   @Before
   public void setUp() throws Exception {
-    factory = new CoreLibraryRuntimePathFactory();
+    RuntimePathBuilder builder = new RuntimePathBuilder();
+    factory = builder::buildModeSketchPath;
     testMode = RuntimePathFactoryTestUtil.createTestJavaMode();
     testImports = RuntimePathFactoryTestUtil.createTestImports();
     testSketch = RuntimePathFactoryTestUtil.createTestSketch();
 
-    classPath = factory.buildClasspath(testMode, testImports, testSketch);
+    classpath = factory.buildClasspath(testMode, testImports, testSketch);
   }
 
   @Test
-  public void testBuildClasspathSize() {
-    assertEquals(2, classPath.size());
+  public void buildClasspathLength() {
+    assertEquals(3, classpath.size());
   }
 
   @Test
-  public void testBuildClasspathValue() {
-    assertEquals("library1", classPath.get(0));
-    assertEquals("library2", classPath.get(1));
+  public void buildClasspathValues() {
+    assertEquals("library6", classpath.get(0));
+    assertEquals("javax.library7", classpath.get(1));
+    assertEquals("library8", classpath.get(2));
   }
-
 
 }
