+++ java/test/processing/mode/java/pdex/util/RuntimePathFactoryTestUtil.java
@@ -18,7 +18,7 @@ along with this program; if not, write to the Free Software Foundation, Inc.
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 */
 
-package processing.mode.java.pdex.util.runtime.strategy;
+package processing.mode.java.pdex.util;
 
 import org.mockito.Mockito;
 import processing.app.Library;
