--- java/test/processing/mode/java/pdex/util/runtime/strategy/ModeSearchRuntimePathFactoryTest.java
@@ -1,63 +1,0 @@
-/* -*- mode: java; c-basic-offset: 2; indent-tabs-mode: nil -*- */
-
-/*
-Part of the Processing project - http://processing.org
-Copyright (c) 2019 The Processing Foundation
-
-This program is free software; you can redistribute it and/or modify
-it under the terms of the GNU General Public License version 2
-as published by the Free Software Foundation.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program; if not, write to the Free Software Foundation, Inc.
-51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
-*/
-
-package processing.mode.java.pdex.util.runtime.strategy;
-
-import org.junit.Before;
-import org.junit.Test;
-import processing.app.Sketch;
-import processing.mode.java.JavaMode;
-import processing.mode.java.pdex.ImportStatement;
-
-import java.util.List;
-
-import static org.junit.Assert.*;
-
-public class ModeSearchRuntimePathFactoryTest {
-
-  private ModeSearchRuntimePathFactory factory;
-  private JavaMode testMode;
-  private List<ImportStatement> testImports;
-  private Sketch testSketch;
-  private List<String> classpath;
-
-  @Before
-  public void setUp() throws Exception {
-    factory = new ModeSearchRuntimePathFactory();
-    testMode = RuntimePathFactoryTestUtil.createTestJavaMode();
-    testImports = RuntimePathFactoryTestUtil.createTestImports();
-    testSketch = RuntimePathFactoryTestUtil.createTestSketch();
-
-    classpath = factory.buildClasspath(testMode, testImports, testSketch);
-  }
-
-  @Test
-  public void buildClasspathLength() {
-    assertEquals(3, classpath.size());
-  }
-
-  @Test
-  public void buildClasspathValues() {
-    assertEquals("library6", classpath.get(0));
-    assertEquals("javax.library7", classpath.get(1));
-    assertEquals("library8", classpath.get(2));
-  }
-
-}
